scripts/requirements.txt
black==23.11.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Deprecated==1.2.14
idna==3.4
junit-xml==1.9
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
